@@ -18,12 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
+  flutter_localizations:
+    sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -31,7 +32,7 @@ dependencies:
   sqflite: ^2.0.0+3
   path_provider: ^2.0.1
   rxdart: ^0.26.0
-  intl: ^0.17.0
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -40,8 +41,8 @@ dev_dependencies:
 flutter_icons:
   android: "launcher_icon"
   ios: true
-  image_path: "assets/icons/nota.png"
-  adaptive_icon_background: "assets/icons/nota.png"
+  image_path: "assets/icons/logo_transparent.png"
+  adaptive_icon_background: "assets/icons/logo_transparent.png"
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
