@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:notes_app/src/blocs/notes_bloc.dart';
 import 'package:notes_app/src/blocs/provider.dart';
+import 'package:notes_app/src/minimal-I18n/app_localizations.dart';
 import 'package:notes_app/src/models/note_model.dart';
 
 class NewNotePage extends StatefulWidget {
@@ -11,17 +12,20 @@ class NewNotePage extends StatefulWidget {
 class _NewNotePageState extends State<NewNotePage> {
   final formKey = GlobalKey<FormState>();
   NoteModel model = NoteModel();
-  NotesBloc notesBloc;
+  late NotesBloc notesBloc;
   @override
   Widget build(BuildContext context) {
     notesBloc = Provider.notesBloc(context);
-    final NoteModel noteData = ModalRoute.of(context).settings.arguments;
+    final NoteModel? noteData =
+        ModalRoute.of(context)!.settings.arguments as NoteModel?;
     if (noteData != null) {
       model = noteData;
     }
     return Scaffold(
       appBar: AppBar(
-        title: Text((noteData == null) ? 'Nueva nota' : model.title),
+        title: Text((noteData == null)
+            ? AppLocalizations.of(context).newNoteTitle
+            : model.title!),
         centerTitle: true,
       ),
       body: SingleChildScrollView(
@@ -44,7 +48,8 @@ class _NewNotePageState extends State<NewNotePage> {
 
   Widget _createHeader() {
     return TextFormField(
-      decoration: InputDecoration(labelText: 'Título'),
+      decoration:
+          InputDecoration(labelText: AppLocalizations.of(context).titleLabel),
       initialValue: (model.title == null) ? "" : model.title,
       textCapitalization: TextCapitalization.sentences,
       onSaved: (value) => model.title = value,
@@ -53,7 +58,8 @@ class _NewNotePageState extends State<NewNotePage> {
 
   Widget _createBody() {
     return TextFormField(
-      decoration: InputDecoration(labelText: 'Nota'),
+      decoration:
+          InputDecoration(labelText: AppLocalizations.of(context).noteLabel),
       initialValue: (model.content == null) ? "" : model.content,
       textCapitalization: TextCapitalization.sentences,
       maxLength: null,
@@ -64,9 +70,9 @@ class _NewNotePageState extends State<NewNotePage> {
   }
 
   void _submit() {
-    if (!formKey.currentState.validate()) return;
+    if (!formKey.currentState!.validate()) return;
     //Dispara el save de todos los textFields del form
-    formKey.currentState.save();
+    formKey.currentState!.save();
     if (model.id == null) {
       notesBloc.createNote(model);
     } else {
