@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:notes_app/src/blocs/provider.dart';
+import 'package:notes_app/src/minimal-I18n/app_localizations.dart';
+import 'package:notes_app/src/minimal-I18n/app_localizations_delegate.dart';
 import 'package:notes_app/src/pages/home_page.dart';
 import 'package:notes_app/src/pages/new_note_page.dart';
 
@@ -17,12 +20,19 @@ class MyApp extends StatelessWidget {
         darkTheme: ThemeData.dark(),
         themeMode: ThemeMode.system,
         debugShowCheckedModeBanner: false,
-        title: 'App de notas',
+        onGenerateTitle: (BuildContext context) =>
+            AppLocalizations.of(context).appTitle,
         initialRoute: 'home',
         routes: {
           'home': (BuildContext context) => HomePage(),
           'new_note': (BuildContext context) => NewNotePage(),
         },
+        localizationsDelegates: [
+          AppLocalizationsDelegate(),
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+        ],
+        supportedLocales: [Locale('en', ''), Locale('es', '')],
       ),
     );
   }
