@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:notes_app/src/blocs/notes_bloc.dart';
 import 'package:notes_app/src/blocs/provider.dart';
+import 'package:notes_app/src/minimal-I18n/app_localizations.dart';
 import 'package:notes_app/src/models/note_model.dart';
 import 'package:notes_app/src/search/search_delegate.dart';
 import 'package:notes_app/src/widgets/note_widget.dart';
@@ -11,7 +12,7 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  List<NoteModel> _notes = [];
+  List<NoteModel>? _notes = [];
 
   @override
   Widget build(BuildContext context) {
@@ -19,7 +20,7 @@ class _HomePageState extends State<HomePage> {
     notesBloc.loadNotes();
     return Scaffold(
       appBar: AppBar(
-        title: Text('Notas'),
+        title: Text(AppLocalizations.of(context).homeTitle),
         centerTitle: true,
         actions: [
           IconButton(
@@ -47,19 +48,19 @@ class _HomePageState extends State<HomePage> {
           if (snapshot.hasData) {
             this._notes = snapshot.data;
 
-            return _notes.isEmpty
+            return _notes!.isEmpty
                 ? ListView.builder(
-                    itemCount: _notes.length,
+                    itemCount: _notes!.length,
                     itemBuilder: (context, i) =>
-                        _createItem(_notes[i], notesBloc, i))
+                        _createItem(_notes![i], notesBloc, i))
                 : GridView.builder(
-                    itemCount: _notes.length,
+                    itemCount: _notes!.length,
                     shrinkWrap: true,
                     gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                         crossAxisCount:
                             (orientation == Orientation.portrait) ? 2 : 3),
                     itemBuilder: (context, i) =>
-                        _createItem(_notes[i], notesBloc, i));
+                        _createItem(_notes![i], notesBloc, i));
           }
           return Center(
             child: CircularProgressIndicator(),
@@ -80,7 +81,7 @@ class _HomePageState extends State<HomePage> {
   void deleteNote(NotesBloc notesBloc, NoteModel note, int index) {
     notesBloc.deleteNote(note).then((value) => {
           setState(() {
-            _notes.removeAt(index);
+            _notes!.removeAt(index);
           })
         });
   }
